Added generalised axis line
The axline() function can now be used to call horizontal (with 'y' parameter), vertical (with 'x' parameter) or diagonal (with 'm' and 'c' parameters) lines. Currently, one can call only an array of x's individually or y's individually or m's and c's together to avoid ambiguity in deciding which line_pars are associated with which line.