Testing kwarg usage on axline()
Added a **kwargs argument to `splotch.axline` which acts as a catch-all for parameters that are not specified explicitly as a splotch `axline()` parameter but will be parsed through to `Line2D` when the matplotlib plot function is called. This `kwargs` dictionary functions identically to the `plot_kw` dictionary (previously called `plot_par`) but the two are combined with those specified in kwargs receiving precedence wherever duplicates exist.